+++ src/main/java/project/repositories/BookingRepository.java
@@ -8,11 +8,12 @@ import org.springframework.data.repository.query.Param;
 
 import project.models.entities.Booking;
 
-public interface BookingRepository extends JpaRepository<Booking, Long> {
+public interface BookingRepository extends JpaRepository<Booking, Integer> {
 	
 	@Query("SELECT b FROM Booking b WHERE b.room.id = :id")
-	List<Booking> findByRoomId(@Param("id") Long id);
+	List<Booking> findByRoomId(@Param("id") int id);
 	
 	@Query("SELECT b FROM Booking b WHERE b.client.id = :clientId AND b.room.id = :roomId")
-	List<Booking> findByClientAndRoom(@Param("clientId") Long clientId, @Param("roomId") Long roomId);
+	List<Booking> findByClientAndRoom(@Param("clientId") int clientId, @Param("roomId") int roomId);
+	
 }
