+++ src/main/java/project/repositories/PhotoRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import project.models.entities.Photo;
 import project.models.entities.Room;
 
-public interface PhotoRepository extends JpaRepository<Photo, Long> {
+public interface PhotoRepository extends JpaRepository<Photo, Integer> {
 	
 	public List<Photo> findByRoom(Room room);
